fix: move ensureDefaultWallets out of route file
Next.js 15 forbids named exports other than HTTP verbs in route files.
moved to src/lib/wallet-helpers.ts.
